Pedagogic solutions

+++ typecheck-perf.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
-agda "src/Examples.agda"
-rm "src/Examples.agdai"
-time agda "src/Examples.agda"
+agda "src/Polynomial/Simple/AlmostCommutativeRing.agda"
+agda "src/Polynomial/Simple/Reflection.agda"
+[ -f "src/Examples.agdai" ] && rm "src/Examples.agdai"
+time agda --no-syntactic-equality "src/Examples.agda"
